[block-device] Make the unit tests reusable across crates
